Changed GPIO to mirror C++
