+++ src/services/ADC.py
@@ -16,44 +16,18 @@ class ADC:
         ADC_RES_14BITS = 0x00000004
         ADC_RES_12BITS = 0x00000008
         ADC_RES_10BITS = 0x0000000C
-    class Instance:
-        Resolution: 'ADC.ADCResolution' = None
-        is_on: bool = False
 
     def __init__(self, pin: Pinout):
         self._pin = SharedMemory.get_pin(pin, memory.PinType.ADC)
-        self.Instance.is_on = False
-        self._pin.data.is_on = False
         
-    def turn_on(self):
-        self._pin.data.is_on = True
-        self.Instance.is_on = True
+    def get_is_on(self) -> bool:
+        return self._pin.is_on
     
-    def get_value(self)->float:
-        if(self.Instance.is_on):
-            match self.Instance.Resolution:
-                case ADC.ADC_RES_16BITS:
-                    return (self._pin.data.value * self.ADC_MAX_VOLTAGE) / float(self.MAX_16BIT)
-                case ADC.ADC_RES_14BITS:
-                    return (self._pin.data.value * self.ADC_MAX_VOLTAGE) / float(self.MAX_14BIT)
-                case ADC.ADC_RES_12BITS:
-                    return (self._pin.data.value * self.ADC_MAX_VOLTAGE) / float(self.MAX_12BIT)
-                case ADC.ADC_RES_10BITS:
-                    return (self._pin.data.value * self.ADC_MAX_VOLTAGE) / float(self.MAX_10BIT)
-                case _:
-                    return 0
-                
-    def get_int_value(self) -> int:
-        if(self.Instance.is_on):
-            match self.Instance.Resolution:
-                case ADC.ADC_RES_16BITS:
-                    return (self._pin.data.value)
-                case ADC.ADC_RES_14BITS:
-                    return (self._pin.data.value << 2)
-                case ADC.ADC_RES_12BITS:
-                    return (self._pin.data.value <<4)
-                case ADC.ADC_RES_10BITS:
-                    return (self._pin.data.value << 6)
-                case _:
-                    return 0
+    def set_value(self, value: int):
+        if(value <= ADC.ADC_MAX_VOLTAGE):
+            self._pin.value = value
+        else:
+            raise ValueError("Value is greater than the maximum voltage")
+    
+
